+++ src/constants/models.ts
@@ -1,0 +1,4 @@
+import { Rank, File } from 'src/types';
+
+export const ranks: Rank[] = [1, 2, 3, 4, 5, 6, 7, 8];
+export const files: File[] = ['A', 'B', 'C', 'D', 'E', 'F', 'G', 'H'];
